Rename module ogn.aprs_utils to 'ogn.parser.utils'
