+++ ogn/model/aircraft_beacon.py
@@ -2,7 +2,7 @@ import re
 
 from sqlalchemy import Column, String, Integer, Float, Boolean, SmallInteger
 
-from ogn.aprs_utils import fpm2ms
+from ogn.parser.utils import fpm2ms
 from .beacon import Beacon
 from ogn.exceptions import OgnParseError
