+++ ogn/model/beacon.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from sqlalchemy import Column, String, Integer, Float, DateTime
 from sqlalchemy.ext.declarative import AbstractConcreteBase
 
-from ogn.aprs_utils import createTimestamp, dmsToDeg, kts2kmh, feet2m
+from ogn.parser.utils import createTimestamp, dmsToDeg, kts2kmh, feet2m
 from ogn.exceptions import AprsParseError
 from .base import Base
