+++ tests/model/test_aircraft_beacon.py
@@ -2,7 +2,7 @@ import unittest
 
 from datetime import datetime
 
-from ogn.aprs_utils import ms2fpm
+from ogn.parser.utils import ms2fpm
 from ogn.model import Beacon, AircraftBeacon
 from ogn.exceptions import OgnParseError
