+++ tests/model/test_beacon.py
@@ -2,7 +2,7 @@ import unittest
 
 from datetime import datetime
 
-from ogn.aprs_utils import dmsToDeg, kts2kmh, m2feet
+from ogn.parser.utils import dmsToDeg, kts2kmh, m2feet
 from ogn.model import Beacon
 from ogn.exceptions import AprsParseError
